google-cloud-functions/requirements.txt
functions-framework
google-cloud-error-reporting
google-cloud-bigquery
google-cloud-storage
requests
python-dotenv